tests - rename "support" directory to simplify
